requirements.txt
Flask
PyPDF2
google-generativeai
langchain-google-genai
langchain
faiss-cpu
python-dotenv
langchain-community
